% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_retention_stats.R
\name{info_retention_stats}
\alias{info_retention_stats}
\title{Information retention summary statistics}
\usage{
info_retention_stats(scaling_result, class_name_field)
}
\arguments{
\item{scaling_result}{data frame with scaling results returned by function mdgp_scale}

\item{class_name_field}{column name in the scaling results data frame that contains the scaled class names}
}
\value{
A list with two data frames. First data frame contains class-specific cell class name (class_name), class frequencies (freq), class raster id (class_id), class proportion (prop), information retention mean (inf_retention_mn) and standard deviations (inf_retention_sd). The second data frame contains the mean (mean) and standard deviation (sd) for the scaled landscape across all classes.
}
\description{
Information retention summary statistics
}
\examples{
# load categorical raster
r <- terra::rast(system.file("extdata/nlm_mid_geom_r3_sa0.tif", package = "landscapeScaling"))

# subset the raster to the lower 300 by 300 pixels
r_sub <- terra::crop(r,terra::ext(0,300,0,300))

# generate relative abundance for the scaled grid
rel_abund <- relative_abundance_scaled_grid(r_sub,class_field='cover',scale_factor=15)

head(rel_abund)

# classify relative abundance samples to multi-dimensional grid points
mdgp_result <- mdgp_scale(rel_abund,parts=3,rpr_threshold=10,monotypic_threshold=90)

head(mdgp_result)

# generate class-specific and landscape scale information retention statistics
infRetStats <- info_retention_stats(mdgp_result,'class_name')

print(infRetStats)
}
