% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscapeScaling-package.R
\docType{package}
\name{landscapeScaling-package}
\alias{landscapeScaling}
\alias{landscapeScaling-package}
\title{landscapeScaling: Scaling Algorithms for Categorical Raster Data}
\description{
Methods to scale categorical raster data. The purpose of the “landscapeScaling” package is to provide a method and the supporting functions to upscale categorical raster data. The recommended method is the multi-dimensional grid-point (mdgp) scaling algorithm. This algorithm generates a new classification scheme on the basis of user desired class label precision of mixed classes and representativeness of the scaled class across the landscape of interest. The scaling output provides scaled categorical raster maps with mixed classes, a corresponding continuous raster with information retention calculations for each scaled grid cell and class-specific and landscape-scale mean and standard deviation of information retention.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gannd/landscapeScaling}
  \item \url{https://gannd.github.io/landscapeScaling/}
  \item Report bugs at \url{https://github.com/gannd/landscapeScaling/issues}
}

}
\author{
\strong{Maintainer}: Daniel Gann \email{gannd@fiu.edu} (\href{https://orcid.org/0000-0001-6131-3391}{ORCID})

}
\keyword{internal}
