% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdgp_generate.R
\name{mdgp_generate}
\alias{mdgp_generate}
\title{Generate multi-dimensional grid points}
\usage{
mdgp_generate(parts = 2, richness = 2)
}
\arguments{
\item{parts}{integer indicating the number of partitions (precision of class-label names)}

\item{richness}{the number of classes (categories in categorical data)}
}
\value{
A data frame with all possible combinations of percentages (rows) for the requested number of partitions and richness (number of columns = richness).
}
\description{
Generate multi-dimensional grid points
}
\examples{
mdgp_generate(parts=4,richness=3)
}
