% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_abundance_scaled_grid.R
\name{relative_abundance_scaled_grid}
\alias{relative_abundance_scaled_grid}
\title{Generate relative class abundances of original (finer resolution) grid cells for scaled (coarser resolution) grid cells}
\usage{
relative_abundance_scaled_grid(x, class_field, scale_factor, verbose = FALSE)
}
\arguments{
\item{x}{SpatRaster object with categorical data (factor)}

\item{class_field}{field name in the attribute table of the raster to be used as class names}

\item{scale_factor}{ratio of lower (scaled) resolution to higher (original) raster resolution (resolution of lower resolution divided by resolution of higher resolution)}

\item{verbose}{bolean; if TRUE progress will be printed in console; default = FALSE}
}
\value{
A data frame with relative abundance of classes. The first two columns are x and y coordinates; all other columns are counts of cells for each class.
}
\description{
Generate relative class abundances of original (finer resolution) grid cells for scaled (coarser resolution) grid cells
}
\examples{
# load categorical raster
r <- terra::rast(system.file("extdata/nlm_mid_geom_r3_sa0.tif", package = "landscapeScaling"))

# subset the raster
r_sub <- terra::crop(r,terra::ext(0,300,0,300))

# generate relative abundance of original classes for a scaled grid with scale factor of 15
rel_abund <- relative_abundance_scaled_grid(r_sub,class_field='cover',scale_factor=15)

head(rel_abund)
}
