% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_result_to_raster.R
\name{scaling_result_to_raster}
\alias{scaling_result_to_raster}
\title{Convert scaling result to spatial raster}
\usage{
scaling_result_to_raster(
  scaling_result,
  class_name_field,
  proj_info = NA,
  scale_factor
)
}
\arguments{
\item{scaling_result}{data frame with scaling results generated by function mdgp_scale}

\item{class_name_field}{field name that contains scaled class names}

\item{proj_info}{projection information of coordinates in the data frame to be converted to raster}

\item{scale_factor}{ratio of lower (scaled) resolution to higher (original) raster resolution (resolution of lower resolution divided by resolution of higher resolution)}
}
\value{
List with 3 objects. Two SpatRaster objects and a list with two objects.\cr
The two scaled raster results are a categorical SpatRaster object with the scaled classified map and scaled class labels in the attribute table, and a numeric SpatRaster object with cell-level information retention. The second object contains a data frame with class-specific information retention and landscape-scale information retention.
}
\description{
converts the scaling result of mdgp_scale() to spatial raster objects#'
}
\examples{
# load categorical raster
r <- terra::rast(system.file("extdata/nlm_mid_geom_r3_sa0.tif", package = "landscapeScaling"))

# subset the raster to the lower 300 by 300 pixels
r_sub <- terra::crop(r,terra::ext(0,300,0,300))

# generate relative abundance for the scaled grid
rel_abund <- relative_abundance_scaled_grid(r_sub,class_field='cover',scale_factor=15)

head(rel_abund)

# classify relative abundance samples to multidimensional grid points
mdgp_result <- mdgp_scale(rel_abund,parts=3,rpr_threshold=10,monotypic_threshold=90)

head(mdgp_result)

# rasterize the scaling result
scaled_map <- scaling_result_to_raster(mdgp_result,class_name_field='class_name',scale_factor=15)

print(scaled_map)

# plot the scaled raster
# scaled color scheme for six lasses
clr_scale <- c('#E69F00','#56B4E9','#009E73','#F0E442','#0072B2','#D55E00')
terra::plot(scaled_map[[1]],col=clr_scale,mar=c(1.5,1.5,1,1))

# plot the information retention raster
terra::plot(scaled_map[[2]],plg=list(ext=c(310,315,20,220),loc = "right"),
col=gray.colors(20,start=0.1,end=1),mar=c(1.5,1.5,1,1))

}
\seealso{
\code{\link{relative_abundance_scaled_grid}} to generate relative abundance for each scaled grid cell,
\code{\link{mdgp_scale}} classifying relative abundance samples to multi-dimensional grid points.
}
